builder/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "builder"
version = "0.1.0"
description = "A Python library for building conversational AI systems"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["ai", "conversational", "intent", "nlp", "chatbot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]



dependencies = [
    "openai>=1.0.0",
    "requests>=2.25.0",
    "pydantic>=1.8.0",
    "typing-extensions>=4.0.0",
    "langchain>=0.1.0",
    "psycopg2-binary>=2.9.0",
    "pandas>=1.3.0",
    "python-dotenv>=0.19.0",
    "sqlalchemy>=1.4.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["builder_package*", "builder*"]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.0.0",
    "black>=21.0.0",
    "flake8>=3.8.0",
    "pyright>=1.1.0",
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.pyright_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=builder",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.pyright]
pythonVersion = "3.8"
typeCheckingMode = "strict"
reportMissingImports = true
reportMissingTypeStubs = false
reportUnusedImport = true
reportPrivateImportUsage = true
reportUnnecessaryTypeIgnoreComment = true
